moved listeners, made lots of stuff protected isntead of public, moved alien
