+++ src/main/java/io/github/addoncommunity/galactifun/base/aliens/TitanAlien.java
@@ -1,6 +1,6 @@
 package io.github.addoncommunity.galactifun.base.aliens;
 
-import io.github.addoncommunity.galactifun.api.alien.Alien;
+import io.github.addoncommunity.galactifun.api.universe.world.Alien;
 import org.bukkit.entity.EntityType;
 import org.bukkit.event.entity.EntityDeathEvent;
